Move existing rules tests into rules module
